pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zzgz-autoto-core"
version = "1.0.0"
description = "多平台自动化核心库，为小红书、微信公众号等平台发布提供基础能力"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "zzgz"}
]
keywords = ["automation", "core", "library", "wechat", "xhs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "playwright",
    "pillow",
    "requests",
]

[project.urls]
Homepage = "https://github.com/henryczq/zzgz-autoto-core"

[tool.setuptools.packages.find]
where = ["."]
include = ["zzgz_autoto_core*"]
exclude = ["zzgz_autoto_core.browser_data*", "zzgz_autoto_core/.browser_data*"]
